requirements.txt
pandas==2.2.2
numpy==1.23.2
pysqlite3==0.5.2
FuzzyTM>=0.4.0
cython>=0.29.21
blosc2==2.0.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.2
